refactor(fieldmap): extract print logic to a dedicated utility
